requirements.txt
boto3
Flask
flask-accepts
Flask-Cors
flask-marshmallow
Flask-Migrate
flask-restx
Flask_SQLAlchemy
Flask-Whooshee
Flask_WTF
marshmallow
marshmallow-sqlalchemy
psycopg2
PyJWT
PyPDF2
pytest
pytest-cov
pytz
requests
requests-oauthlib
WeasyPrint
WTForms